replace float with double (better precision)

+++ include/candidate.hpp
@@ -17,7 +17,7 @@
 __global__ 
 void candidatePreComputation(
     const uchar* F, size_t Fstep,
-    float* S, size_t Sstep,
+    double* S, size_t Sstep,
     int* D, size_t Dstep,
     int width, int height
 );
@@ -25,7 +25,7 @@ void candidatePreComputation(
 
 __global__ 
 void candidateThresholdKernel(
-    const float *S, size_t Sstep,
+    const double *S, size_t Sstep,
     const int *D, size_t Dstep,
     uchar *C, size_t Cstep,
     int width, int height
