+++ include/iterative.hpp
@@ -12,22 +12,22 @@
 
 
 __host__ __device__
-thrust::tuple<float,float,int> upgradeCandidate(
+thrust::tuple<double,double,int> upgradeCandidate(
     const uchar* F, size_t Fstep,
-    thrust::tuple<float,float,int> cand,
+    thrust::tuple<double,double,int> cand,
     int width, int height
 );
 
 __host__
-std::vector<std::tuple<float,int,int>> sortThresholdCandidates(
-    const float *S, size_t Sstep, 
+std::vector<std::tuple<double,int,int>> sortThresholdCandidates(
+    const double *S, size_t Sstep, 
     int width, int height
 );
 
 __host__
 cv::Mat candidateIterativeSearch(
     const uchar* F, size_t Fstep,
-    const float *S, size_t Sstep,
+    const double *S, size_t Sstep,
     const int *D, size_t Dstep,
     int width, int height
 );
@@ -37,7 +37,7 @@ __host__
 void candidateExpand(
     const uchar *F, size_t Fstep, 
     uchar *CI, size_t CIstep, 
-    float startY, float startX,
+    double startY, double startX,
     int dir, 
     int width, int height
 );
@@ -47,7 +47,7 @@ __host__
 bool setCandidates(
     const uchar *F, size_t Fstep,
     uchar *CI, size_t CIstep, 
-    float y, float x,
+    double y, double x,
     int dir, 
     int width, int height
 );
