+++ src/color.cu
@@ -5,31 +5,31 @@ __host__ __device__
 thrust::tuple<uchar,uchar,uchar> bicubicInterpolation(
     const uchar* F,
     size_t Fstep,
-    float y, float x,
+    double y, double x,
     int width, int height
 ) {
     // clamp‑to‑edge strategy for out of range subpixels
     int x0 = static_cast<int>(floor(x));
     int y0 = static_cast<int>(floor(y));
-    float dx = x - x0;
-    float dy = y - y0;
+    double dx = x - x0;
+    double dy = y - y0;
 
     // atmull-Rom spline, a = -0.5
-    auto cubicWeight = [] __host__ __device__ (float t) {
-        const float a = -0.5;
+    auto cubicWeight = [] __host__ __device__ (double t) {
+        const double a = -0.5;
         t = fabs(t);
-        if (t < 1.0f) return (a+2)*t*t*t - (a+3)*t*t + 1;
+        if (t < 1.0) return (a+2)*t*t*t - (a+3)*t*t + 1;
         else if (t < 2.0) return a*t*t*t - 5*a*t*t + 8*a*t - 4*a;
-        else return 0.0f;
+        else return 0.0;
     };
 
-    float wx[4], wy[4];
+    double wx[4], wy[4];
     for (int i=0; i<4; i++) {
         wx[i] = cubicWeight(dx - (i-1));
         wy[i] = cubicWeight(dy - (i-1));
     }
 
-    float C[3] = {0, 0, 0};
+    double C[3] = {0, 0, 0};
     for (int j=0; j<4; j++) {
         for (int i=0; i<4; i++) {
             int xi = min(max(x0+i-1,0), width-1);
@@ -42,7 +42,7 @@ thrust::tuple<uchar,uchar,uchar> bicubicInterpolation(
             uchar c1 = pix.y;
             uchar c2 = pix.z;
             //
-            float wxy = wx[i]*wy[j];
+            double wxy = wx[i]*wy[j];
             C[0] += c0 * wxy;
             C[1] += c1 * wxy;
             C[2] += c2 * wxy;
@@ -60,7 +60,7 @@ __host__ __device__
 thrust::tuple<uchar,uchar,uchar> getColorChannels(
     const uchar* F,
     size_t Fstep,
-    float y, float x,
+    double y, double x,
     int width, int height
 ) {
     int rY = round(y);
