+++ src/score.cu
@@ -1,28 +1,28 @@
 #include "score.hpp"
 
 __host__ __device__
-float getRad(int direction) {
+double getRad(int direction) {
     direction %= DIRECTIONS;
     return direction*(PI / DIRECTIONS);
 }
 
 __host__ __device__ 
-thrust::tuple<float,float> getUnitVector(int dir) { // y x
-    float rad = getRad(dir);
+thrust::tuple<double,double> getUnitVector(int dir) { // y x
+    double rad = getRad(dir);
     while (rad >= PI) rad -= PI;
     return thrust::make_tuple(sin(rad), cos(rad));
 }
 
 __host__ __device__ 
-thrust::tuple<float,float> getOrthogonalUnitVector(int dir) { // y x
+thrust::tuple<double,double> getOrthogonalUnitVector(int dir) { // y x
     dir = (dir + DIRECTIONS/2) % DIRECTIONS;
     return getUnitVector(dir);
 }
 
 __host__ __device__ 
-void insertionSort(float* a, int n) {
+void insertionSort(double* a, int n) {
     for (int i = 1; i < n; ++i) {
-        float key = a[i];
+        double key = a[i];
         int j = i - 1;
         while (j >= 0 && a[j] > key) {
             a[j + 1] = a[j];
@@ -33,19 +33,19 @@ void insertionSort(float* a, int n) {
 }
 
 __host__ __device__
-float emd(const float* arr, int dir) {
+double emd(const double* arr, int dir) {
     //
     int edge = (dir + DIRECTIONS/2) % DIRECTIONS;
     // array sum (in order to normalize)
-    float sum = 0;
+    double sum = 0;
     for (int k = 0; k < 2*DIRECTIONS; k++) {
         sum += arr[k];
     }
     // prefix sum (normalized values)
-    float fixed = (1.0/DIRECTIONS); 
-    float prefixSum1[2*DIRECTIONS], prefixSum2[2*DIRECTIONS];
+    double fixed = (1.0/DIRECTIONS); 
+    double prefixSum1[2*DIRECTIONS], prefixSum2[2*DIRECTIONS];
     for (int k = 0; k < 2*DIRECTIONS; k++) {
-        float val1, val2;
+        double val1, val2;
         if (k % DIRECTIONS == edge) {
             val1 = val2 = fixed/2; 
         } else if (k < edge || k > DIRECTIONS + edge) {
@@ -57,9 +57,9 @@ float emd(const float* arr, int dir) {
         }
         // std::cout << val1 << " | " << val2 << std::endl;
         // prefix sums of delta of the normalized array values 
-        float arrVal = arr[k]/sum;
-        float delta1 = arrVal - val1;
-        float delta2 = arrVal - val2;
+        double arrVal = arr[k]/sum;
+        double delta1 = arrVal - val1;
+        double delta2 = arrVal - val2;
         if (!k) {
             prefixSum1[k] = delta1;
             prefixSum2[k] = delta2;
@@ -72,10 +72,10 @@ float emd(const float* arr, int dir) {
     // median computation
     insertionSort(prefixSum1, 2*DIRECTIONS);
     insertionSort(prefixSum2, 2*DIRECTIONS);
-    float med1 = (prefixSum1[DIRECTIONS - 1] + prefixSum1[DIRECTIONS])/2;
-    float med2 = (prefixSum2[DIRECTIONS - 1] + prefixSum2[DIRECTIONS])/2;
+    double med1 = (prefixSum1[DIRECTIONS - 1] + prefixSum1[DIRECTIONS])/2;
+    double med2 = (prefixSum2[DIRECTIONS - 1] + prefixSum2[DIRECTIONS])/2;
     // ring EMD
-    float emd1 = 0.0f, emd2 = 0.0f;
+    double emd1 = 0.0, emd2 = 0.0;
     for (int k = 0; k < 2*DIRECTIONS; k++) {
         emd1 += fabsf(prefixSum1[k] - med1);
         emd2 += fabsf(prefixSum2[k] - med2);
@@ -85,28 +85,28 @@ float emd(const float* arr, int dir) {
 }
 
 __host__ __device__
-float computeLabScore(
+double computeLabScore(
     const uchar* F,
     size_t Fstep,
-    float yPixel, float xPixel,
+    double yPixel, double xPixel,
     int dir, 
     int width, int height
 ) {
-    float emdSum = 0.0f;
+    double emdSum = 0.0;
     for (int c = 1; c <= CIRCLE_COUNT; c++) {
         //
         int minL = 255, minA = 255, minB = 255;
         for (int d = 0; d < DIRECTIONS; d++) { 
             // 
-            thrust::tuple<float,float> unit = getOrthogonalUnitVector(d);
-            float dY = thrust::get<0>(unit);
-            float dX = thrust::get<1>(unit);
+            thrust::tuple<double,double> unit = getOrthogonalUnitVector(d);
+            double dY = thrust::get<0>(unit);
+            double dX = thrust::get<1>(unit);
             //
-            float y1 = yPixel + c*dY;
-            float x1 = xPixel + c*dX;
+            double y1 = yPixel + c*dY;
+            double x1 = xPixel + c*dX;
             //
-            float y2 = yPixel - c*dY;
-            float x2 = xPixel - c*dX;
+            double y2 = yPixel - c*dY;
+            double x2 = xPixel - c*dX;
             //
             thrust::tuple<uchar,uchar,uchar> lab1 = getColorChannels(F, Fstep, y1, x1, width, height);
             int l1 = thrust::get<0>(lab1);
@@ -123,20 +123,20 @@ float computeLabScore(
             minB = min(minB, min(b1, b2));
         }
         //
-        float lArr[2*DIRECTIONS];
-        float aArr[2*DIRECTIONS];
-        float bArr[2*DIRECTIONS];
+        double lArr[2*DIRECTIONS];
+        double aArr[2*DIRECTIONS];
+        double bArr[2*DIRECTIONS];
         for (int d = 0; d < DIRECTIONS; d++) { 
             // 
-            thrust::tuple<float,float> unit = getOrthogonalUnitVector(d);
-            float dY = thrust::get<0>(unit);
-            float dX = thrust::get<1>(unit);
+            thrust::tuple<double,double> unit = getOrthogonalUnitVector(d);
+            double dY = thrust::get<0>(unit);
+            double dX = thrust::get<1>(unit);
             //
-            float y1 = yPixel + c*dY;
-            float x1 = xPixel + c*dX;
+            double y1 = yPixel + c*dY;
+            double x1 = xPixel + c*dX;
             //
-            float y2 = yPixel - c*dY;
-            float x2 = xPixel - c*dX;
+            double y2 = yPixel - c*dY;
+            double x2 = xPixel - c*dX;
             //
             thrust::tuple<uchar,uchar,uchar> lab1 = getColorChannels(F, Fstep, y1, x1, width, height);
             int l1 = thrust::get<0>(lab1);
@@ -161,23 +161,23 @@ float computeLabScore(
         emdSum += min(emd(lArr, dir), min(emd(bArr, dir), emd(aArr, dir)));
     }
     // compute the EMD-score
-    float maxEmd = 1.0f*DIRECTIONS/4;
-    float emdAv = emdSum/CIRCLE_COUNT;
-    float emdNorm = fminf(emdAv/maxEmd, 1.0f);
-    float emdScore = 1.0 - powf(emdNorm, SCORE_BOOSTER); 
+    double maxEmd = 1.0*DIRECTIONS/4;
+    double emdAv = emdSum/CIRCLE_COUNT;
+    double emdNorm = fminf(emdAv/maxEmd, 1.0);
+    double emdScore = 1.0 - powf(emdNorm, SCORE_BOOSTER); 
     return emdScore;
 }
 
 __host__ __device__
-thrust::tuple<float,int> bestPossibleScore(
+thrust::tuple<double,int> bestPossibleScore(
     const uchar* F, size_t Fstep,
-    float yPixel, float xPixel,
+    double yPixel, double xPixel,
     int width, int height
 ) {
-    float bestScore = -1.0f;
+    double bestScore = -1.0;
     int bestDir = 0;
     for (int d = 0; d < DIRECTIONS; d++) {
-        float score = computeLabScore(F, Fstep, yPixel, xPixel, d, width, height);
+        double score = computeLabScore(F, Fstep, yPixel, xPixel, d, width, height);
         if (score > bestScore) {
             bestScore = score;
             bestDir = d;
